+++ lib/taql.rb
@@ -5,8 +5,8 @@ require_relative "taql/version"
 module Taql
   class Error < StandardError; end
 
-  def self.execute(query)
-    ActiveRecord::Base.connection.execute(query).tap do |result|
+  def self.execute(query, connection: ActiveRecord::Base.connection)
+    connection.execute(query).tap do |result|
       if (results = result.entries).any?
         $stdout.puts Table.new(results)
       end